+++ src/pages/Facility/services/inventory/SupplyDeliveryTable.tsx
@@ -25,6 +25,7 @@ import {
   DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu";
 import { cn } from "@/lib/utils";
+import { calculateTotal } from "@/pages/Facility/services/inventory/externalSupply/utils/inventoryUtils";
 import { MonetaryComponentType } from "@/types/base/monetaryComponent/monetaryComponent";
 import { DeliveryOrderStatus } from "@/types/inventory/deliveryOrder/deliveryOrder";
 import {
@@ -37,52 +38,6 @@ import supplyDeliveryApi from "@/types/inventory/supplyDelivery/supplyDeliveryAp
 import { ShortcutBadge } from "@/Utils/keyboardShortcutComponents";
 import mutate from "@/Utils/request/mutate";
 import { EllipsisVertical } from "lucide-react";
-
-function calculateLineTotal(
-  delivery: SupplyDeliveryRead,
-  internal: boolean,
-): number {
-  const priceComponents = internal
-    ? delivery.supplied_inventory_item?.product?.charge_item_definition
-        ?.price_components
-    : delivery.supplied_item?.charge_item_definition?.price_components;
-
-  if (!priceComponents) return 0;
-
-  const baseComponent = priceComponents.find(
-    (c) => c.monetary_component_type === MonetaryComponentType.base,
-  );
-  const basePrice = baseComponent?.amount
-    ? parseFloat(baseComponent.amount)
-    : 0;
-  const quantity = delivery.supplied_item_quantity || 1;
-
-  let total = basePrice * quantity;
-
-  // Apply taxes
-  priceComponents
-    .filter((c) => c.monetary_component_type === MonetaryComponentType.tax)
-    .forEach((tax) => {
-      if (tax.factor) {
-        total += basePrice * quantity * (tax.factor / 100);
-      } else if (tax.amount) {
-        total += parseFloat(tax.amount);
-      }
-    });
-
-  // Apply discounts
-  priceComponents
-    .filter((c) => c.monetary_component_type === MonetaryComponentType.discount)
-    .forEach((discount) => {
-      if (discount.factor) {
-        total -= basePrice * quantity * (discount.factor / 100);
-      } else if (discount.amount) {
-        total -= parseFloat(discount.amount);
-      }
-    });
-
-  return total;
-}
 
 interface SupplyDeliveryTableProps {
   deliveries: SupplyDeliveryRead[];
@@ -270,7 +225,7 @@ export function SupplyDeliveryTable({
             </TableCell>
             <TableCell>
               <MonetaryDisplay
-                amount={calculateLineTotal(delivery, internal).toFixed(2)}
+                amount={calculateTotal(delivery, internal).toFixed(2)}
               />
             </TableCell>
             <TableCell>
