+++ src/pages/Facility/services/inventory/externalSupply/deliveryOrder/DeliveryOrderShow.tsx
@@ -44,9 +44,12 @@ import { Label } from "@/components/ui/label";
 import { MonetaryDisplay } from "@/components/ui/monetary-display";
 import { RadioGroup, RadioGroupItem } from "@/components/ui/radio-group";
 import { AddSupplyDeliveryForm } from "@/pages/Facility/services/inventory/externalSupply/deliveryOrder/AddSupplyDeliveryForm";
+import {
+  calculateTotal,
+  getInventoryBasePath,
+} from "@/pages/Facility/services/inventory/externalSupply/utils/inventoryUtils";
 import { ProductKnowledgeSelect } from "@/pages/Facility/services/inventory/ProductKnowledgeSelect";
 import { SupplyDeliveryTable } from "@/pages/Facility/services/inventory/SupplyDeliveryTable";
-import { MonetaryComponentType } from "@/types/base/monetaryComponent/monetaryComponent";
 import {
   DELIVERY_ORDER_STATUS_COLORS,
   DeliveryOrderRetrieve,
@@ -65,58 +68,12 @@ import { ShortcutBadge } from "@/Utils/keyboardShortcutComponents";
 import mutate from "@/Utils/request/mutate";
 import query from "@/Utils/request/query";
 
-function calculateDeliveryTotal(
-  delivery: SupplyDeliveryRead,
-  internal: boolean,
-): number {
-  const priceComponents = internal
-    ? delivery.supplied_inventory_item?.product?.charge_item_definition
-        ?.price_components
-    : delivery.supplied_item?.charge_item_definition?.price_components;
-
-  if (!priceComponents) return 0;
-
-  const baseComponent = priceComponents.find(
-    (c) => c.monetary_component_type === MonetaryComponentType.base,
-  );
-  const basePrice = baseComponent?.amount
-    ? parseFloat(baseComponent.amount)
-    : 0;
-  const quantity = delivery.supplied_item_quantity || 1;
-
-  let total = basePrice * quantity;
-
-  // Apply taxes
-  priceComponents
-    .filter((c) => c.monetary_component_type === MonetaryComponentType.tax)
-    .forEach((tax) => {
-      if (tax.factor) {
-        total += basePrice * quantity * (tax.factor / 100);
-      } else if (tax.amount) {
-        total += parseFloat(tax.amount);
-      }
-    });
-
-  // Apply discounts
-  priceComponents
-    .filter((c) => c.monetary_component_type === MonetaryComponentType.discount)
-    .forEach((discount) => {
-      if (discount.factor) {
-        total -= basePrice * quantity * (discount.factor / 100);
-      } else if (discount.amount) {
-        total -= parseFloat(discount.amount);
-      }
-    });
-
-  return total;
-}
-
 function calculateTotalPrice(
   deliveries: SupplyDeliveryRead[],
   internal: boolean,
 ): number {
   return deliveries.reduce(
-    (sum, delivery) => sum + calculateDeliveryTotal(delivery, internal),
+    (sum, delivery) => sum + calculateTotal(delivery, internal),
     0,
   );
 }
@@ -445,7 +402,18 @@ export function DeliveryOrderShow({
       <div className="space-y-6">
         <div className="flex flex-col sm:flex-row sm:items-center justify-between gap-2">
           <div className="flex items-center gap-4">
-            <BackButton size="icon" className="shrink-0">
+            <BackButton
+              size="icon"
+              className="shrink-0"
+              to={getInventoryBasePath(
+                facilityId,
+                locationId,
+                internal,
+                false,
+                isRequester,
+                "",
+              )}
+            >
               <ChevronLeft />
             </BackButton>
             <div>
